Renames streamquery to scheduled_queries
