+++ streamalert/scheduled_queries/__init__.py
