+++ streamalert/scheduled_queries/container/container.py
