+++ streamalert/scheduled_queries/handlers/athena.py
