+++ streamalert/scheduled_queries/query_packs/packs/athena.py
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from streamalert.streamquery.query_packs.configuration import QueryPackConfiguration
+from streamalert.scheduled_queries.query_packs.configuration import QueryPackConfiguration
 
 QueryPackConfiguration(
     name='athena_any_query',
