+++ streamalert/scheduled_queries/state/state_manager.py
