+++ streamalert/scheduled_queries/support/clock.py
