+++ terraform/modules/tf_scheduled_queries/variables.tf
