+++ terraform/modules/tf_scheduled_queries/versions.tf
