+++ tests/unit/streamalert/scheduled_queries/handlers/athena_test.py
@@ -18,7 +18,7 @@ from dateutil.tz import tzlocal
 from mock import MagicMock
 from nose.tools import assert_equals, assert_true, assert_false
 
-from streamalert.streamquery.handlers.athena import (
+from streamalert.scheduled_queries.handlers.athena import (
     AthenaClient,
     AthenaQueryExecution,
     AthenaQueryResult
