+++ tests/unit/streamalert/scheduled_queries/query_packs/manager_test.py
@@ -17,7 +17,7 @@ from datetime import datetime, timezone
 from mock import MagicMock
 from nose.tools import assert_equals, assert_false, assert_true
 
-from streamalert.streamquery.query_packs.manager import (
+from streamalert.scheduled_queries.query_packs.manager import (
     QueryPack, QueryPacksManager, QueryParameterGenerator
 )
