+++ tests/unit/streamalert/scheduled_queries/state/.gitignore
